% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.all.schedule.R
\name{cpm.all.schedule}
\alias{cpm.all.schedule}
\alias{cpm_all_schedule}
\alias{genAllCpmSched}
\title{Generates all minimum makespan schedules for a cpm network}
\usage{
cpm.all.schedule(cpm)
}
\arguments{
\item{vector}{with early start time for all activities}

\item{vector}{with slack for all activities}
}
\value{
Matrix with all mininum makespan cpm schedules (start time for all activities)
}
\description{
Generates all minimum makespan schedules for a cpm network
}
\examples{

ex.cpm.activities.duration <- c(1,4,5,7,2,3,1)
ex.cpm.activities.successors <- list(c(2,3), 4, c(4,5), 6, 7, 7, c(0))
ex.cpm <- cpm(ex.cpm.activities.duration, 
              ex.cpm.activities.successors)

#ex.cpm is a list:
# - est (Early Start Time)  - ex.cpm$est
# - eft (Early Finish Time) - ex.cpm$eft
# - lst (Late Start Time)   - ex.cpm$lst
# - lft (Late Finish Time)  - ex.cpm$lft

ex.cpm.activities.schedule <- cpm.all.schedule(ex.cpm)
                                               
}
\seealso{
Other scheduling: \code{\link{cpm}},
  \code{\link{mmf.all.sequences}},
  \code{\link{mmf.get.breakeven}},
  \code{\link{mmf.get.selffunding}}, \code{\link{mmf.npv}}
}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}
