% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\alias{critical.path.method}
\alias{critical_path_method}
\title{The Critical Path Method (CPM) is a scheduling algorithm that produces the  
minimum makespan schedule for a project with unlimited resources. As a byproduct, it generates
the slacks for the non-critical activities. 
Inputs: activity precedence relation, activity duration relation  
Output: early and late start and finish time for all project activities}
\usage{
cpm(activities.duration = c(1, 4, 5, 7, 2, 3, 1),
  activities.successors = list(c(2, 3), 4, c(4, 5), 6, 7, 7, c(0)))
}
\arguments{
\item{activities.duration:}{vector with the duration of the project activities.}

\item{activities.successors:}{list with the set of successors for each activity.}
}
\value{
Returns a list with 4 vectors: (1) EST (Early Start Time), (2) EFT(Early Finish Time), (3) LST(Lately Start Time), 
(4) LFT (Lately Finish Time)
}
\description{
The Critical Path Method (CPM) is an algorithm that generates the minimum makespan
schedule for a given project activity network with unlimited resources. 

CPM generates also the list of non-critical activities with their respective slack.
}
\examples{
ex.cpm.activities.duration <- c(1,4,5,7,2,3,1)
ex.cpm.activities.successors <- list(c(2,3), 4, c(4,5), 6, 7, 7, c(0))
ex.cpm <- cpm(ex.cpm.activities.duration, 
              ex.cpm.activities.successors)

}
\seealso{
Other scheduling: \code{\link{cpm.all.schedule}},
  \code{\link{mmf.all.sequences}},
  \code{\link{mmf.get.breakeven}},
  \code{\link{mmf.get.selffunding}}, \code{\link{mmf.npv}}
}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}
