% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.path.method.R
\name{critical.path.method}
\alias{cpm}
\alias{critical.path.method}
\alias{critical_path_method}
\title{Scheduling a set of project activities}
\usage{
critical.path.method(activities.duration, activities.successors)
}
\arguments{
\item{activities.duration}{Vector with activities duration.}

\item{activities.successors}{Vector with dependencies between activities.}
}
\description{
The Critical Path Method or Critical Path Analysis, is a 
mathematically based algorithm for scheduling a set of project activities. 

CPM will get how long your complex project will take to complete and which 
activities are "critical," meaning that they have to be done on time or else 
the whole project will take longer.

CPM calculates: 

As input, CPM will receive: 

   - A list of all activities required to complete the project;
   
   - The time (duration) that each activity will take to completion;
   
   - The dependencies between the activities. 
   
   
And will return:

   - The optimized sequence of activities that must be performed to guarantee
   the shorthest duration.
}
\examples{
ex.cpm.activities.duration <- c(1,4,5,7,2,3,1)
ex.cpm.activities.successors <- list(c(2,3), 4, c(4,5), 6, 7, 7, c(0))
ex.cpm <- critical.path.method(ex.cpm.activities.duration, 
                               ex.cpm.activities.successors)

}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}

