% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounted.csf.R
\name{discounted.csf}
\alias{discounted.csf}
\alias{discountedCsf}
\alias{discounted_csf}
\title{The cash flows incomes/outcomes applying the Tax Rate to the present time.}
\usage{
discounted.csf(cfs = c(-350, 100, 200, 150, 75), interest.rate = 0.0619,
  begin.of.period = FALSE)
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{interest.rate}{A number that represents the nominal Interest Rate, 
presented by year.}

\item{begin.of.period}{A boolean that represents if the Tax Rate will be 
applied at the begining of period. FALSE by default, the Tax Rate will be 
applied to the second period.}
}
\value{
The vector of cash flows incomes/outcomes applying the Tax Rate to
the present time.
}
\description{
The cash flows incomes/outcomes applying the Tax Rate to the present time.
}
\examples{

ex.disc.csf <- discounted.csf(c(-350,100,200,150,75), 0.0619, FALSE)

}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{draw.cfs}}, \code{\link{ifir}},
  \code{\link{net.future.value}},
  \code{\link{net.present.value}}
}
\keyword{cash}
\keyword{flow}
\keyword{series}

