% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.discounted.cash.R
\name{draw.discounted.cash}
\alias{draw.discounted.cash}
\alias{draw_discounted_cash}
\title{Draw Discounted Cash vs Time}
\usage{
draw.discounted.cash(discounted.cash)
}
\arguments{
\item{discounted.cash}{A vector with discounted cash flow for each timestamp.}
}
\description{
Draw Discounted Cash vs Time
}

