% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.graph.R
\name{draw.graph}
\alias{draw.graph}
\alias{drawGraph}
\alias{draw_graph}
\title{Draw the graph imported from the spreadsheet.}
\usage{
draw.graph(edges = c(1, 2, 1, 3, 2, 3))
}
\arguments{
\item{edges}{- A vector defining the edges, the first edge points from the 
first element to the second, the second edge from the third to the fourth, 
etc.}
}
\value{
graph.image.path - The path to the generated graph file.
}
\description{
This function is responsible for ploting the graph based on the 
edges and export the image to a file.
}
\examples{

ex.graph.image.path <- draw.graph(c(1,2, 1,3, 2,3, 3,4))

}
\keyword{draw}
\keyword{drawGraph}
\keyword{graph}

