% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.list.to.xls.R
\name{excel.list.to.xls}
\alias{excel.list.to.xls}
\alias{excelListToXls}
\alias{excel_list_to_xls}
\title{Export the generated ifm package results to a spreadsheet.}
\arguments{
\item{list.ifm.result}{The list with all results proccessed by the IFM 
package.}
}
\value{
file.path The path to the generated file.
}
\description{
This function is responsible for reading a list of objects and
export a spreadsheet with the results proccessed by the IFM package.
The file contains:
\itemize{
 \item The raw data frame used to calculate the maxNPV, minSF and minBKE;
 \item The image oh the generated graph;
 \item The image oh the "Discounted Cash x Time" chart;
 \item The image oh the "MPV (ca$h) x Schedulling ID" chart;
 \item The image oh the "Self Funding (time) x Schedulling ID" chart;
 \item The image oh the "Breaking Event (time) x Schedulling ID" chart;
}
}
\examples{

ex.file.path <- excel.list.to.xls(list.ifm.result)

}
\seealso{
Other utility: \code{\link{excel.xls.to.list}}
}
\keyword{excel}
\keyword{list.to.xls}
\keyword{maxNPV}

