% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.xls.to.list.R
\name{excel.xls.to.list}
\alias{excel.xls.to.list}
\alias{excelXlsToList}
\alias{excel_xls_to_list}
\title{Extract a list of variables from the spreadsheet to be used on the maxNPV
 function.}
\arguments{
\item{xls.spreadsheet.path}{The complete path to the spreadsheet that
represents the project.}
}
\value{
List of variables to be used on the maxNPV function.
}
\description{
This function is responsible for reading a spreadsheet 
representing the project, and return a list with the following information
(in this order): The interest rate, the list of activities, the list of 
durations of activities, the list of predecessors of activities and the 
matrix that represents the cash flow series
}
\examples{

list.of.variables <- excel.xls.to.list("C://temp//spreadsheet.xlsx")

}
\keyword{excel}
\keyword{maxNPV}
\keyword{xls.to.list}

