% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future.value.R
\name{future.value}
\alias{future.value}
\alias{futureValue}
\alias{future_value}
\title{Calculate the future value of an asset at a specific date.}
\usage{
future.value(present.value, interest.rate, number.of.periods)
}
\arguments{
\item{present.value}{A number that represents the present value of the money.}

\item{interest.rate}{A number that represents the interest rate.}

\item{number.of.periods}{A number that represent the number of periods.}
}
\description{
It measures the nominal future sum of money that a given sum of 
money is "worth" at a specified time in the future assuming a certain 
interest rate, or more generally, rate of return.
}
\examples{

ex.fv <- future.value(1000, 1.1425, 12)

}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{draw.cfs}},
  \code{\link{inflation.free.interest.rate}},
  \code{\link{net.future.value}},
  \code{\link{net.present.value}},
  \code{\link{present.value}}
}
\keyword{futureValue}

