\name{ifm-package}
\alias{ifm-package}
\alias{ifm}
\docType{package}
\title{
\packageTitle{ifm}
}
\description{
\packageDescription{ifm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ifm}
\packageIndices{ifm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ifm}

Maintainer: \packageMaintainer{ifm}
}
\references{
~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%% ~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## examples here...
}