\name{imf-package}
\alias{imf-package}
\alias{imf}
\docType{package}
\title{
\packageTitle{imf}
}
\description{
\packageDescription{imf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imf}
\packageIndices{imf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{imf}

Maintainer: \packageMaintainer{imf}
}
\references{
~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%% ~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## examples here...
}