% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.free.interest.rate.R
\name{inflation.free.interest.rate}
\alias{IFIR}
\alias{IfIR}
\alias{ifir}
\alias{inflation.free.interest.rate}
\title{Calculate the Inflation-free Interest Rate.}
\usage{
inflation.free.interest.rate(interest.rate = 14.25, inflation.rate = 7.59)
}
\arguments{
\item{interest.rate}{A number that represents the nominal Interest Rate, presented by year.}

\item{inflation.rate}{A number that represents the Inflation Rate, presented by year.}
}
\description{
Calculate the Inflation-free Interest Rate.
}
\examples{

ex.ifir <- inflation.free.interest.rate(14.25, 12)

}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{draw.cfs}}, \code{\link{future.value}},
  \code{\link{present.value}}
}
\keyword{inflation-free,}
\keyword{interest}
\keyword{rate}

