% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.df.1r.R
\name{mmf.df.1r}
\alias{mmf.df.1r}
\title{Generates a data frame with Sequence, Schedule, NPV, Breakeven and 
Self Funding}
\usage{
mmf.df.1r(mmf.seq, mmf.sched, mmf.npv)
}
\arguments{
\item{mmf.seq}{A list of sequences}

\item{mmf.sched}{A list of schedules}

\item{mmf.nvp}{A list of NPV values}
}
\description{
Generates a data frame with Sequence, Schedule, NPV, Breakeven and 
Self Funding
}

