% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.df.infr.R
\name{mmf.df.infr}
\alias{mmf.df.infr}
\title{Generates a data frame with Schedule, NPV, Breakeven and Self Funding}
\usage{
mmf.df.infr(mmf.sched, mmf.npv, mmf.npv.selffunding, mmf.npv.breakeven)
}
\arguments{
\item{mmf.sched}{A list of schedules}

\item{mmf.npv}{A list of NPV values}

\item{mmf.npv.selffunding}{A list of Selffunding times}

\item{mmf.npv.breakeven}{A list of Breakeven times}
}
\description{
Generates a data frame with Schedule, NPV, Breakeven and Self Funding
}
\examples{

ex.sheet.data <- excel.xls.to.list("resources/spreadsheet.xls")
ex.sheet.data.interest.rate <- ex.sheet.data[[1]]
ex.sheet.data.activities <- ex.sheet.data[[2]]
ex.sheet.data.durations <- ex.sheet.data[[3]]
ex.sheet.data.predecessors <- ex.sheet.data[[4]]
ex.sheet.data.cfs <- ex.sheet.data[[5]]

ex.mmf.df.infr <- mmf.df.infr(ex.cpm.schedules,
                          ex.cpm.npv,
                          ex.cpm.npv.selffunding,
                          ex.cpm.npv.breakeven)
                          
}
