% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.max.npv.R
\name{mmf.max.npv}
\alias{mmf.max.npv}
\alias{mmf_max_npv}
\title{Calculates NPV for all schedules}
\usage{
mmf.max.npv(cfs, durations, all.sequences, interest.rate)
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{durations}{A vector with a list of activities durations.}

\item{all.sequences}{List of all possible MMF sequences.}

\item{interest.rate}{A number that represents the nominal Interest Rate, 
presented by year.}
}
\value{
One sequence with max NPV and its value
}
\description{
Calculates NPV for all schedules
}
\examples{
 
ex.sheet.data <- excel.xls.to.list("../resources/spreadsheet.xls")
ex.sheet.data.interest.rate <- ex.sheet.data[[1]]
ex.sheet.data.activities <- ex.sheet.data[[2]]
ex.sheet.data.durations <- ex.sheet.data[[3]]
ex.sheet.data.predecessors <- ex.sheet.data[[4]]
ex.sheet.data.cfs <- ex.sheet.data[[5]]
 
ex.mmf.seq <- mmf.all.sequences(ex.sheet.data.predecessors)
 
ex.mmf.npv <- mmf.max.npv(ex.sheet.data.cfs,
                          ex.sheet.data.durations,
                          ex.mmf.seq,
                          ex.sheet.data.interest.rate)
 
ex.mmf.npv.max <- which.max(ex.mmf.npv)


}
\seealso{
Other scheduling: \code{\link{cpm.all.schedule}},
  \code{\link{critical.path.method}},
  \code{\link{mmf.all.sequences}}
}
\keyword{activities}
\keyword{features,}
\keyword{marketable}
\keyword{minimum}
\keyword{project}
\keyword{scheduling,}

