% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.future.value.R
\name{net.future.value}
\alias{net.future.value}
\alias{nfv}
\title{Net Future Value is a combination of different future values from different 
times, all which are put into one larger present value.}
\usage{
net.future.value(cfs, interest.rate, begin.of.period = FALSE)
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{interest.rate}{A number that represents the nominal Interest Rate, 
presented by year.}

\item{begin.of.period}{A boolean that represents if the Tax Rate will be 
applied at the begining of period. FALSE by default.}
}
\value{
A vector with values updated to future value.
}
\description{
Net Future Value is a combination of different future values 
from different times, all which are put into one larger present value.
}
\examples{

ex.nfv <- net.future.value(c(-350,100,200,150,75), 6.19, TRUE)

}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{draw.cfs}}, \code{\link{future.value}},
  \code{\link{inflation.free.interest.rate}},
  \code{\link{net.present.value}},
  \code{\link{present.value}}
}
\keyword{futureValue}

