% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predecessors.to.edges.R
\name{predecessors.to.edges}
\alias{predecessors.to.edges}
\alias{predecessorsToEdges}
\alias{predecessors_to_edges}
\title{Generate a vector with the edges to be plotted by draw.graph function.}
\usage{
predecessors.to.edges(list.of.predecessors = list())
}
\arguments{
\item{list.of.predecessors}{The vector that contain all activities predecessors.}
}
\value{
edges - List of edges to be used on the draw.graph function.
}
\description{
This function is responsible for reading a vector with all 
activities predecessors and generate a list of the edges that will be 
plotted by the draw.graph function.
}
\examples{

ex.edges <- predecessors.to.edges(ex.sheet.data$predecessors)

}
\seealso{
Other utility: \code{\link{excel.list.to.xls}},
  \code{\link{excel.xls.to.list}}
}
\keyword{draw}
\keyword{draw.graph}

