% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.value.R
\name{present.value}
\alias{present.value}
\alias{presentValue}
\alias{present_value}
\title{Calculate the present value of an asset at a specific date.}
\usage{
present.value(future.value, interest.rate, number.of.periods)
}
\arguments{
\item{future.value}{A number that represents the future value of the money.}

\item{interest.rate}{A number that represents the interest rate.}

\item{number.of.periods}{A number that represent the number of periods.}
}
\description{
In economics, present value, also known as present discounted 
value, is the value of an expected income stream determined as of the date 
of valuation. The present value is always less than or equal to the future 
value because money has interest-earning potential, a characteristic referred
to as the time value of money, except during times of negative interest 
rates, when the present value will be less than the future value.
}
\examples{

ex.pv <- present.value(1000, 1.1425, 12)
print(ex.pv)

}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{draw.cfs}}, \code{\link{future.value}},
  \code{\link{inflation.free.interest.rate}},
  \code{\link{net.present.value}}
}
\keyword{presentValue}

