% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.all.schedule.R
\name{cpm.all.schedule}
\alias{cpm.all.schedule}
\alias{cpm_all_schedule}
\alias{genAllCpmSched}
\title{Generates all minimum makespan schedules for a cpm network}
\usage{
cpm.all.schedule(cpm)
}
\arguments{
\item{matrix}{with est,eft, lst, lft for all activities}
}
\value{
Matrix with all mininum makespan cpm schedules (start time for all activities)
}
\description{
Generates all minimum makespan schedules for a cpm network
}
\examples{
 
#ex.cpm is the list generated by the cpm() function:
# - est (Early Start Time)  - ex.cpm$est
# - eft (Early Finish Time) - ex.cpm$eft
# - lst (Late Start Time)   - ex.cpm$lst
# - lft (Late Finish Time)  - ex.cpm$lft
ex.cpm<-cpm()
ex.cpm.activities.schedule <- cpm.all.schedule(ex.cpm)
                                               
}
\seealso{
Other scheduling: \code{\link{cpm}},
  \code{\link{mmf.all.sequences}},
  \code{\link{mmf.get.breakeven}},
  \code{\link{mmf.get.selffunding}}, \code{\link{mmf.npv}}
}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}
