% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\alias{critical.path.method}
\alias{critical_path_method}
\title{The critical path method (CPM) is a step-by-step project management technique for process 
planning that defines critical and non-critical tasks with the goal of preventing time-frame
problems and process bottlenecks#' activities are "critical," meaning that they have to be done on time or else 
the whole project will take longer.}
\usage{
cpm(activities.duration = c(1, 4, 5, 7, 2, 3, 1),
  activities.successors = list(c(2, 3), 4, c(4, 5), 6, 7, 7, c(0)))
}
\arguments{
\item{activities.duration}{Vector with activities duration.}

\item{activities.successors}{Vector with dependencies between activities.}
}
\value{
Returns list of EST (Early Start Time), EFT(Early Finish Time),LST(Lately Start Time), LFT (Lately Finish Time) using Forward Pass and Backward Pass
}
\description{
The Critical Path Method or Critical Path Analysis, is a 
mathematically based algorithm for scheduling a set of project activities. 

CPM will get how long your complex project will take to complete and which 
activities are "critical," meaning that they have to be done on time or else 
the whole project will take longer.
}
\examples{
ex.cpm.activities.duration <- c(1,4,5,7,2,3,1)
ex.cpm.activities.successors <- list(c(2,3), 4, c(4,5), 6, 7, 7, c(0))
ex.cpm <- cpm(ex.cpm.activities.duration, 
              ex.cpm.activities.successors)

}
\seealso{
Other scheduling: \code{\link{cpm.all.schedule}},
  \code{\link{mmf.all.sequences}},
  \code{\link{mmf.get.breakeven}},
  \code{\link{mmf.get.selffunding}}, \code{\link{mmf.npv}}
}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}

