% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.cfs.R
\name{draw.cfs}
\alias{draw.cfs}
\alias{drawCfs}
\alias{draw_cfs}
\title{Draw the graph of cash flow.}
\usage{
draw.cfs(cfs, gt = "Cash Flow Graphic")
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{gt}{A title for the graph.}
}
\description{
Draw the graph of cash flow.
}
\examples{

ex.cfs <- c(-2000,1000,1500,-500,500)
draw.cfs(ex.cfs,'My Cash Flow')
}
\seealso{
Other financial: \code{\link{discount.rate.vector}},
  \code{\link{future.value}},
  \code{\link{inflation.free.interest.rate}},
  \code{\link{net.present.value}},
  \code{\link{present.value}}
}
\keyword{drawcfs}

