% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.all.sequences.R
\name{mmf.all.sequences}
\alias{mmf.all.sequences}
\alias{mmf_all_sequences}
\title{Generates the list of all possible MMF sequences (topsorts), constrained by the predecessors}
\usage{
mmf.all.sequences(predecessors = 0)
}
\arguments{
\item{predecessors}{List of Predecessors - Zero for none.  The index of the list of predecessors represents the id of MMF and the value
in that position the id of MMF's predecessors.  For instance, in list(0,1,2,3,1,5,c(4,6)), we have for MMF 1, predecessor MMF 0; for MMF 2,
predecessor MMF 7, predecessors 4 and 6}
}
\value{
List of all possible MMF sequences.
}
\description{
Generates the list of all possible MMF sequences (topsorts), constrained by the predecessors
}
\examples{

ex.activities.predecessors<-list(0,1,2,3,1,5,c(4,6))
ex.mmf.seq <- mmf.all.sequences(ex.activities.predecessors)

}
\seealso{
Other scheduling: \code{\link{cpm.all.schedule}},
  \code{\link{critical.path.method}}, \code{\link{mmf.npv}}
}
\keyword{activities}
\keyword{features,}
\keyword{marketable}
\keyword{minimum}
\keyword{project}
\keyword{scheduling,}

