% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.max.npv.R
\name{mmf.max.npv}
\alias{mmf.max.npv}
\title{Calculates NPV for all schedules}
\usage{
mmf.max.npv()
}
\description{
@name mmf.max.npv
 
 @aliases mmf_max_npv
 
 @return One sequence with max NPV and its value
 
 @export mmf.max.npv
 
 @keywords minimum marketable features, scheduling, project activities
 
 @family scheduling
 
 @example
 
 ex.max.npv <- mmf.max.npv(c(-100, 80, 81, 82, 83, 84, 85, 86), 
             c(1, 4, 5, 7, 2, 3, 1),
             c(1,2,3,4,5,6,7),
             0.024)
}

