% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.max.npv.R
\name{mmf.max.npv}
\alias{mmf.max.npv}
\title{Return Max NPV}
\usage{
mmf.max.npv(mmf.npv, mmf.seq, mmf.schedules)
}
\arguments{
\item{mmf.npv}{Vector of Net Present Value}

\item{mmf.seq}{Vector with the sequence of activities}

\item{mmf.schedules}{Vector with the collection of possible schedules}
}
\value{
list with NPV, sequence and scheduleof the sequence with the maximum NPV
}
\description{
this function identifies the sequence of activities and respectivies schedules where with the optimized NPV
}
\examples{

ex.sheet.data <- excel.xls.to.list("resources/spreadsheet.xls")
ex.sheet.data.interest.rate <- ex.sheet.data[[1]]
ex.sheet.data.activities <- ex.sheet.data[[2]]
ex.sheet.data.durations <- ex.sheet.data[[3]]
ex.sheet.data.predecessors <- ex.sheet.data[[4]]
ex.sheet.data.cfs <- ex.sheet.data[[5]]

ex.mmf.seq <- mmf.all.sequences(ex.sheet.data.predecessors)

ex.mmf <- mmf.npv(ex.sheet.data.cfs,
                  ex.sheet.data.durations,
                  ex.mmf.seq,
                  ex.sheet.data.interest.rate)

ex.mmf.shedules <- ex.mmf[['shedules']]
ex.mmf.cfs.nominal <- ex.mmf[['cfs.nominal']]
ex.mmf.cfs.discounted <- ex.mmf[['cfs.discounted']]
ex.mmf.npv <- ex.mmf[['npv']]

}
\keyword{npv}

